+++ apps/examples/simple-example/test/integration/test_it_query_something_extended.py
@@ -5,11 +5,11 @@ from datetime import datetime, timezone
 import pytest  # type: ignore
 
 from hopeit.testing.apps import execute_event
-from hopeit.server.version import ENGINE_VERSION
+from hopeit.server.version import APPS_API_VERSION
 
 from model import Something, SomethingNotFound, Status, StatusType
 
-APP_VERSION = ENGINE_VERSION.replace('.', "x")
+APP_VERSION = APPS_API_VERSION.replace('.', "x")
 
 
 @pytest.fixture